Add handling for access to single AWS Account
When a user only has access to one AWS account, they will not see the
web form with a selection of roles and accounts available to them. This
commit will prompt the user to enter an account name if only one was
found and will not fail with the error "Principal not found" anymore.